Column class unit tests
  * Test all public functions
  * Card value and colour type cast to integer
  * Make `canPlace` method more obvious

+++ src/Card.php
@@ -83,8 +83,8 @@ class Card implements Api\GameObjectInterface
     public function setNumber(int $cardNumber)
     {
         $this->number = $cardNumber;
-        $this->suit   = $suit = $cardNumber % 4;
-        $this->value  = floor($cardNumber / 4);
+        $this->suit   = (int) $suit = $cardNumber % 4;
+        $this->value  = (int) floor($cardNumber / 4);
         $this->colour = (self::SUIT_DIAMOND == $suit || self::SUIT_HEART == $suit)
             ? self::COLOUR_RED
             : self::COLOUR_BLACK;
